refactor: move enhancers to separate folder
